Create README.md
30-Day Python Programming Challenge:

Welcome to my 30-day Python programming challenge! In this challenge, I'll be creating 30 different Python projects, ranging from beginner to advanced levels. The goal of this challenge is to improve my Python skills and knowledge, and to have fun while doing it!

Project Ideas
Here's a list of the project ideas I'll be working on during this challenge:
1. Dice Rolling

How to Use

Each project will have its own folder, containing the source code and any necessary files. Simply navigate to the folder of the project you're interested in, and follow the instructions in the README file.

Contributing
If you have any suggestions or improvements for any of the projects, feel free to open a pull request or submit an issue.


About Me
I'm a self taught Python programmer. I created this challenge to improve my skills and share my knowledge with others. You can find me on LinkedIn and Twitter.